refactor: 상수관련 파일 코드컨벤션에 맞춰 salaryRange 폴더에서 constance 폴더로 이동 및 파일에 수정된 경로 적용
