+++ src/constants/salaryRange.ts
